# doc/openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "/odms/v1"
    }
  ],
  "info": {
    "description": "API for the SENG302 ODMS Web-Client architecture, story 41 for team 100",
    "version": "0.1.4",
    "title": "ODMS API for SENG302"
  },
...
  "paths": {
...
    "/admins": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Gets all the administrators",
        "parameters": [
          {
            "in": "query",
            "name": "startIndex",
            "description": "Number of admins to skip before returning result",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "count",
            "description": "Number of admins to return",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "q",
            "description": "Search string for admin names",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "Admin": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/adminsOverview"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Creates a new admin",
        "security": [
          {
            "Admin": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/adminOverview"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/admins/{username}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "description": "Username of the admin",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Admin"
        ],
        "summary": "Gets the administrator matching the username",
        "security": [
          {
            "Admin": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/adminDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "description": "Username of the admin",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Admin"
        ],
        "summary": "Edits the administrator matching the username",
        "security": [
          {
            "Admin": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/adminOverview"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "description": "Username of the admin",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Admin"
        ],
        "summary": "Deletes the administrator matching the username",
        "security": [
          {
            "Admin": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "date": {
        "type": "object",
        "description": "date object",
        "properties": {
          "year": {
            "type": "integer",
            "maximum": 9999,
            "minimum": 0
          },
          "month": {
            "type": "integer",
            "maximum": 12,
            "minimum": 1
          },
          "day": {
            "type": "integer",
            "maximum": 31,
            "minimum": 1
          }
        }
      },
      "name": {
        "type": "object",
        "description": "Name of the role",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "middleNames": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },
...
      "adminsOverview": {
        "type": "array",
        "items": {
          "type": "object",
          "allOf": [
            {
              "$ref": "#/components/schemas/adminOverview"
            }
          ]
        }
      },
      "adminOverview": {
        "type": "object",
        "allOf": [
          {
            "type": "object",
            "properties": {
              "username": {
                "$ref": "#/components/schemas/username"
              },
              "name": {
                "$ref": "#/components/schemas/name"
              }
            }
          }
        ]
      },
      "adminDetails": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/adminOverview"
          },
          {
            "type": "object",
            "properties": {
              "lastModified": {
                "$ref": "#/components/schemas/datetime"
              },
              "created": {
                "$ref": "#/components/schemas/datetime"
              }
            }
          }
        ]
      },
      "datetime": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/date"
          },
          {
            "type": "object",
            "properties": {
              "hour": {
                "type": "integer",
                "minimum": 0,
                "maximum": 23
              },
              "minute": {
                "type": "integer",
                "minimum": 0,
                "maximum": 59
              },
              "second": {
                "type": "integer",
                "minimum": 0,
                "maximum": 59
              }
            }
          }
        ]
      },
      "username": {
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "User": {
        "type": "apiKey",
        "name": "X-Authorization",
        "in": "header"
      },
      "Clinician": {
        "type": "apiKey",
        "name": "X-Authorization",
        "in": "header"
      },
      "Admin": {
        "type": "apiKey",
        "name": "X-Authorization",
        "in": "header"
      }
    }
  }
}